Move shared data into data.h
